requirements.txt
flask
websocket-client
psycopg2-binary